First Moodle 2.4 version (CHANGES.txt contains pending tasks)

+++ html/moodle2/lib/weblib.php
@@ -713,7 +713,7 @@ class moodle_url {
     public static function make_draftfile_url($draftid, $pathname, $filename, $forcedownload = false) {
         global $CFG, $USER;
         $urlbase = "$CFG->httpswwwroot/draftfile.php";
-        $context = get_context_instance(CONTEXT_USER, $USER->id);
+        $context = context_user::instance($USER->id);
 
         return self::make_file_url($urlbase, "/$context->id/user/draft/$draftid".$pathname.$filename, $forcedownload);
     }
@@ -1072,11 +1072,11 @@ function format_text($text, $format = FORMAT_MOODLE, $options = NULL, $courseid_
         if (is_object($options['context'])) {
             $context = $options['context'];
         } else {
-            $context = get_context_instance_by_id($options['context']);
+            $context = context::instance_by_id($options['context']);
         }
     } else if ($courseid_do_not_use) {
         // legacy courseid
-        $context = get_context_instance(CONTEXT_COURSE, $courseid_do_not_use);
+        $context = context_course::instance($courseid_do_not_use);
     } else {
         // fallback to $PAGE->context this may be problematic in CLI and other non-standard pages :-(
         $context = $PAGE->context;
@@ -1227,14 +1227,16 @@ function format_text($text, $format = FORMAT_MOODLE, $options = NULL, $courseid_
 /**
  * Resets all data related to filters, called during upgrade or when filter settings change.
  *
- * @global object
- * @global object
+ * @param bool $phpunitreset true means called from our PHPUnit integration test reset
  * @return void
  */
-function reset_text_filters_cache() {
+function reset_text_filters_cache($phpunitreset = false) {
     global $CFG, $DB;
 
-    $DB->delete_records('cache_text');
+    if (!$phpunitreset) {
+        $DB->delete_records('cache_text');
+    }
+
     $purifdir = $CFG->cachedir.'/htmlpurifier';
     remove_dir($purifdir, true);
 }
@@ -1272,7 +1274,7 @@ function format_string($string, $striplinks = true, $options = NULL) {
 
     if (is_numeric($options)) {
         // legacy courseid usage
-        $options  = array('context'=>get_context_instance(CONTEXT_COURSE, $options));
+        $options  = array('context'=>context_course::instance($options));
     } else {
         $options = (array)$options; // detach object, we can not modify it
     }
@@ -1281,7 +1283,7 @@ function format_string($string, $striplinks = true, $options = NULL) {
         // fallback to $PAGE->context this may be problematic in CLI and other non-standard pages :-(
         $options['context'] = $PAGE->context;
     } else if (is_numeric($options['context'])) {
-        $options['context'] = get_context_instance_by_id($options['context']);
+        $options['context'] = context::instance_by_id($options['context']);
     }
 
     if (!$options['context']) {
@@ -1412,7 +1414,7 @@ function format_text_email($text, $format) {
 function format_module_intro($module, $activity, $cmid, $filter=true) {
     global $CFG;
     require_once("$CFG->libdir/filelib.php");
-    $context = get_context_instance(CONTEXT_MODULE, $cmid);
+    $context = context_module::instance($cmid);
     $options = array('noclean'=>true, 'para'=>false, 'filter'=>$filter, 'context'=>$context, 'overflowdiv'=>true);
     $intro = file_rewrite_pluginfile_urls($activity->intro, 'pluginfile.php', $context->id, 'mod_'.$module, 'intro', null);
     return trim(format_text($intro, $activity->introformat, $options, null));
@@ -1838,7 +1840,26 @@ function get_html_lang($dir = false) {
 
 /**
  * Send the HTTP headers that Moodle requires.
- * @param $cacheable Can this page be cached on back?
+ *
+ * There is a backwards compatibility hack for legacy code
+ * that needs to add custom IE compatibility directive.
+ *
+ * Example:
+ * <code>
+ * if (!isset($CFG->additionalhtmlhead)) {
+ *     $CFG->additionalhtmlhead = '';
+ * }
+ * $CFG->additionalhtmlhead .= '<meta http-equiv="X-UA-Compatible" content="IE=8" />';
+ * header('X-UA-Compatible: IE=8');
+ * echo $OUTPUT->header();
+ * </code>
+ *
+ * Please note the $CFG->additionalhtmlhead alone might not work,
+ * you should send the IE compatibility header() too.
+ *
+ * @param string $contenttype
+ * @param bool $cacheable Can this page be cached on back?
+ * @return void, sends HTTP headers
  */
 function send_headers($contenttype, $cacheable = true) {
     global $CFG;
@@ -1846,6 +1867,10 @@ function send_headers($contenttype, $cacheable = true) {
     @header('Content-Type: ' . $contenttype);
     @header('Content-Script-Type: text/javascript');
     @header('Content-Style-Type: text/css');
+
+    if (empty($CFG->additionalhtmlhead) or stripos($CFG->additionalhtmlhead, 'X-UA-Compatible') === false) {
+        @header('X-UA-Compatible: IE=edge');
+    }
 
     if ($cacheable) {
         // Allow caching on "back" (but not on normal clicks)
@@ -2076,7 +2101,7 @@ function print_group_picture($group, $courseid, $large=false, $return=false, $li
         }
     }
 
-    $context = get_context_instance(CONTEXT_COURSE, $courseid);
+    $context = context_course::instance($courseid);
 
     // If there is no picture, do nothing
     if (!$group->picture) {
@@ -2132,7 +2157,7 @@ function print_recent_activity_note($time, $user, $text, $link, $return=false, $
     $output = '';
 
     if (is_null($viewfullnames)) {
-        $context = get_context_instance(CONTEXT_SYSTEM);
+        $context = context_system::instance();
         $viewfullnames = has_capability('moodle/site:viewfullnames', $context);
     }
 
@@ -2185,7 +2210,8 @@ function navmenulist($course, $sections, $modinfo, $strsection, $strjumpto, $wid
     $menu = array();
     $doneheading = false;
 
-    $coursecontext = get_context_instance(CONTEXT_COURSE, $course->id);
+    $courseformatoptions = course_get_format($course)->get_format_options();
+    $coursecontext = context_course::instance($course->id);
 
     $menu[] = '<ul class="navmenulist"><li class="jumpto section"><span>'.$strjumpto.'</span><ul>';
     foreach ($modinfo->cms as $mod) {
@@ -2194,7 +2220,8 @@ function navmenulist($course, $sections, $modinfo, $strsection, $strjumpto, $wid
             continue;
         }
 
-        if ($mod->sectionnum > $course->numsections) {   /// Don't show excess hidden sections
+        // For course formats using 'numsections' do not show extra sections
+        if (isset($courseformatoptions['numsections']) && $mod->sectionnum > $courseformatoptions['numsections']) {
             break;
         }
 
@@ -2205,8 +2232,9 @@ function navmenulist($course, $sections, $modinfo, $strsection, $strjumpto, $wid
         if ($mod->sectionnum >= 0 and $section != $mod->sectionnum) {
             $thissection = $sections[$mod->sectionnum];
 
-            if ($thissection->visible or !$course->hiddensections or
-                      has_capability('moodle/course:viewhiddensections', $coursecontext)) {
+            if ($thissection->visible or
+                    (isset($courseformatoptions['hiddensections']) and !$courseformatoptions['hiddensections']) or
+                    has_capability('moodle/course:viewhiddensections', $coursecontext)) {
                 $thissection->summary = strip_tags(format_string($thissection->summary,true));
                 if (!$doneheading) {
                     $menu[] = '</ul></li>';
@@ -2551,23 +2579,38 @@ function obfuscate_text($plaintext) {
  * @param string $email The email address to display
  * @param string $label The text to displayed as hyperlink to $email
  * @param boolean $dimmed If true then use css class 'dimmed' for hyperlink
+ * @param string $subject The subject of the email in the mailto link
+ * @param string $body The content of the email in the mailto link
  * @return string The obfuscated mailto link
  */
-function obfuscate_mailto($email, $label='', $dimmed=false) {
+function obfuscate_mailto($email, $label='', $dimmed=false, $subject = '', $body = '') {
 
     if (empty($label)) {
         $label = $email;
     }
-    if ($dimmed) {
-        $title = get_string('emaildisable');
-        $dimmed = ' class="dimmed"';
-    } else {
-        $title = '';
-        $dimmed = '';
+
+    $label = obfuscate_text($label);
+    $email = obfuscate_email($email);
+    $mailto = obfuscate_text('mailto');
+    $url = new moodle_url("mailto:$email");
+    $attrs = array();
+
+    if (!empty($subject)) {
+        $url->param('subject', format_string($subject));
     }
-    return sprintf("<a href=\"%s:%s\" $dimmed title=\"$title\">%s</a>",
-                    obfuscate_text('mailto'), obfuscate_email($email),
-                    obfuscate_text($label));
+    if (!empty($body)) {
+        $url->param('body', format_string($body));
+    }
+
+    // Use the obfuscated mailto
+    $url = preg_replace('/^mailto/', $mailto, $url->out());
+
+    if ($dimmed) {
+        $attrs['title'] = get_string('emaildisable');
+        $attrs['class'] = 'dimmed';
+    }
+
+    return html_writer::link($url, $label, $attrs);
 }
 
 /**
@@ -2814,7 +2857,7 @@ function convert_tabrows_to_tree($tabrows, $selected, $inactive, $activated) {
  * @return bool
  */
 function debugging($message = '', $level = DEBUG_NORMAL, $backtrace = null) {
-    global $CFG, $USER, $UNITTEST;
+    global $CFG, $USER;
 
     $forcedebug = false;
     if (!empty($CFG->debugusers) && $USER) {
@@ -2836,17 +2879,13 @@ function debugging($message = '', $level = DEBUG_NORMAL, $backtrace = null) {
         }
         $from = format_backtrace($backtrace, CLI_SCRIPT);
         if (PHPUNIT_TEST) {
-            echo 'Debugging: ' . $message . "\n" . $from;
+            if (phpunit_util::debugging_triggered($message, $level, $from)) {
+                // We are inside test, the debug message was logged.
+                return true;
+            }
+        }
 
-        } else if (!empty($UNITTEST->running)) {
-            // When the unit tests are running, any call to trigger_error
-            // is intercepted by the test framework and reported as an exception.
-            // Therefore, we cannot use trigger_error during unit tests.
-            // At the same time I do not think we should just discard those messages,
-            // so displaying them on-screen seems like the only option. (MDL-20398)
-            echo '<div class="notifytiny">' . $message . $from . '</div>';
-
-        } else if (NO_DEBUG_DISPLAY) {
+        if (NO_DEBUG_DISPLAY) {
             // script does not want any errors or debugging in output,
             // we send the info to error log instead
             error_log('Debugging: ' . $message . $from);
